Did literally everything
There are now three interfaces: a debug console interface for local
hacking, a slight extension of it for playing via netcat, and a version
using Python Websockets that serves to a Single Page App (included as
the HTML file). The latter has a point-and-click interface.

Game logic has been bugfixed in many ways to approximate the
paper-and-pencil version of the game. There are still some bugs, e.g.,
with transitive cycles.